requirements.txt
Flask==3.1.0
Flask_Compress==1.17
Flask_Sitemap==0.4.0
Flask_SocketIO==5.5.0
pydantic==2.10.5
python-dotenv==1.0.1
